fix(db): store discount value as BIGINT unsigned per the money rule
promos.value and vouchers.value hold a fixed IDR amount in the fixed
discount case, so they are money and must be BIGINT UNSIGNED (§7,
golden rule 5) like min_spend / max_discount already are — they were
unsignedInteger. Widened both; no behavior change within range, just
brings the schema in line with the rule.

+++ database/migrations/2026_06_26_010000_create_promos_table.php
@@ -15,7 +15,9 @@ return new class extends Migration
             $table->id();
             $table->string('code')->unique();
             $table->enum('type', ['percentage', 'fixed']);
-            $table->unsignedInteger('value');
+            // Holds a percentage (0–100) or a fixed IDR amount; money is
+            // BIGINT UNSIGNED per §7 / golden rule 5.
+            $table->unsignedBigInteger('value');
             $table->unsignedBigInteger('max_discount')->nullable();
             $table->unsignedBigInteger('min_spend')->nullable();
             $table->timestamp('expiry_date');
